# LoadEventNexus code cleanup and small performance improvements