--- a/Framework/DataHandling/inc/MantidDataHandling/ProcessBankData.h
+++ b/Framework/DataHandling/inc/MantidDataHandling/ProcessBankData.h
@@ -55,6 +55,7 @@ private:
   size_t getWorkspaceIndexFromPixelID(const detid_t pixID);
   size_t getFirstEventIndex(const size_t pulseIndex) const;
   size_t getLastEventIndex(const size_t pulseIndex, const size_t numPulses) const;
+  void preCountAndReserveMem();
 
   /// Algorithm being run
   DefaultEventLoader &m_loader;