--- a/Framework/DataHandling/src/BankPulseTimes.cpp
+++ b/Framework/DataHandling/src/BankPulseTimes.cpp
@@ -7,6 +7,7 @@
 #include "MantidDataHandling/BankPulseTimes.h"
 
 #include <nexus/NeXusFile.hpp>
+#include <numeric>
 
 using namespace Mantid::Kernel;
 //===============================================================================================
@@ -37,38 +38,55 @@ BankPulseTimes::BankPulseTimes(::NeXus::File &file, const std::vector<int> &pNum
   else
     file.getAttr("offset", startTime);
   Mantid::Types::Core::DateAndTime start(startTime);
-  // Load the seconds offsets
 
-  const auto heldTimeZeroType = file.getInfo().type;
+  // number of pulse times
+  const auto dataInfo = file.getInfo();
+  const int64_t numValues = std::accumulate(dataInfo.dims.cbegin(), dataInfo.dims.cend(), 0,
+                                            [](const int64_t accum, const int64_t value) { return accum + value; });
+  if (numValues == 0)
+    throw std::runtime_error("event_time_zero field has no data!");
+
+  const auto heldTimeZeroType = dataInfo.type;
+
   // Nexus only requires event_time_zero to be a NXNumber, we support two
   // possilites
   if (heldTimeZeroType == ::NeXus::FLOAT64) {
-    std::vector<double> seconds;
-    file.getData(seconds);
-    file.closeData();
-    // Now create the pulseTimes
-    if (seconds.size() == 0)
-      throw std::runtime_error("event_time_zero field has no data!");
-
-    std::transform(seconds.cbegin(), seconds.cend(), std::back_inserter(pulseTimes),
-                   [start](double seconds) { return start + seconds; });
+    this->readData<double>(file, numValues, start);
   } else if (heldTimeZeroType == ::NeXus::UINT64) {
-    std::vector<uint64_t> nanoseconds;
-    file.getData(nanoseconds);
-    file.closeData();
-    // Now create the pulseTimes
-    if (nanoseconds.size() == 0)
-      throw std::runtime_error("event_time_zero field has no data!");
-
-    std::transform(nanoseconds.cbegin(), nanoseconds.cend(), std::back_inserter(pulseTimes),
-                   [start](int64_t nanoseconds) { return start + nanoseconds; });
+    this->readData<uint64_t>(file, numValues, start);
   } else {
     throw std::invalid_argument("Unsupported type for event_time_zero");
   }
+  file.closeData();
+
   // Ensure that we always have a consistency between nPulses and
   // periodNumbers containers
   if (pulseTimes.size() != pNumbers.size()) {
     periodNumbers = std::vector<int>(pulseTimes.size(), FirstPeriod);
+  }
+}
+
+template <typename ValueType>
+void BankPulseTimes::readData(::NeXus::File &file, int64_t numValues, Mantid::Types::Core::DateAndTime &start) {
+  std::vector<int64_t> indexStart{0};
+  std::vector<int64_t> indexStep{std::min(numValues, static_cast<int64_t>(3600 * 60))}; // 1 hour at 60Hz
+
+  // getSlab needs the data allocated already
+  std::vector<ValueType> rawData(indexStep[0]);
+
+  // loop over chunks of data and transform each chunk
+  while ((indexStep[0] > 0) && (indexStart[0] < numValues)) {
+    file.getSlab(rawData.data(), indexStart, indexStep);
+
+    // Now create the pulseTimes
+    std::transform(rawData.cbegin(), rawData.cend(), std::back_inserter(pulseTimes),
+                   [start](ValueType incremental_time) { return start + incremental_time; });
+
+    // increment the slab to get
+    indexStart[0] += indexStep[0];
+    indexStep[0] = std::min(indexStep[0], numValues - indexStart[0]);
+    // resize the vector
+    rawData.resize(indexStep[0]);
   }
 }