@peterfpeterson 2023-11-03
It appears that loading 2.6 milllion pulse times at once is faster than reading it in chunks and accumulating. I increased the minimum size for chunking which got rid of the regression you saw.