@williamfgc 2023-11-04
Beware the above is only CPU profiling. I'd be careful of merging without comprehensive performance testing (as done in #28495 ). The time reported for this PR is worse on SNS analysis, so it might affect other production systems using LoadEventNexus.